+++ SurfaceTopography/ChangeLog.md
@@ -7,6 +7,13 @@ Deprecations in version 1.0
 - `unit` entry for the `info` dictionary will disappear
 - `scale_factor` property for scaled topographies will disappear
 
+v0.100.2 (23Nov21)
+------------------
+
+- API: Add `has_undefined_data` property to `NonuniformLineScan`
+- MAINT: Avoid reading full file in Matlab reader
+- MAINT: Avoid reading full file in OPD reader
+
 v0.100.1 (22Nov21)
 ------------------